+++ src/clj/rems/api/services/command.clj
@@ -50,7 +50,8 @@
    :get-catalogue-item-licenses applications/get-catalogue-item-licenses
    :get-workflow workflow/get-workflow
    :allocate-application-ids! applications/allocate-application-ids!
-   :get-attachment-metadata attachments/get-attachment-metadata})
+   :get-attachment-metadata attachments/get-attachment-metadata
+   :copy-attachment! attachments/copy-attachment!})
 
 (defn command! [cmd]
   ;; Use locks to prevent multiple commands being executed in parallel.